# frontend/src/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "BranchOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "code"
        ],
        "title": "BranchOut",
        "type": "object"
      },
...
      "DepartmentOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "code"
        ],
        "title": "DepartmentOut",
        "type": "object"
      },
      "DevLoginRequest": {
        "properties": {
          "email": {
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "DevLoginRequest",
        "type": "object"
      },
...
      "FeedbackCreate": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "kind": {
            "$ref": "#/components/schemas/FeedbackKind"
          },
          "query_id": {
            "format": "uuid",
            "title": "Query Id",
            "type": "string"
          }
        },
        "required": [
          "query_id",
          "kind"
        ],
        "title": "FeedbackCreate",
        "type": "object"
      },
      "FeedbackKind": {
        "enum": [
          "wrong",
          "outdated",
          "unhelpful",
          "helpful"
        ],
        "title": "FeedbackKind",
        "type": "string"
      },
      "FeedbackOut": {
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "cited": {
            "items": {
              "type": "string"
            },
            "title": "Cited",
            "type": "array"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/FeedbackKind"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "query_id": {
            "format": "uuid",
            "title": "Query Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "reporter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "routed_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routed To"
          },
          "status": {
            "$ref": "#/components/schemas/FeedbackStatus"
          }
        },
        "required": [
          "id",
          "query_id",
          "kind",
          "comment",
          "status",
          "created_at",
          "question",
          "answer",
          "outcome",
          "cited",
          "reporter",
          "routed_to",
          "resolution_note"
        ],
        "title": "FeedbackOut",
        "type": "object"
      },
      "FeedbackPatch": {
        "properties": {
          "resolution_note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "status": {
            "$ref": "#/components/schemas/FeedbackStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "FeedbackPatch",
        "type": "object"
      },
      "FeedbackStatus": {
        "enum": [
          "open",
          "acknowledged",
          "resolved"
        ],
        "title": "FeedbackStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "clinician",
          "author",
          "approver",
          "admin"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "branch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BranchOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "department": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DepartmentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "role",
          "branch",
          "department"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Source-backed answers from approved institutional documents. ProtoCite retrieves and displays approved documents; it does not provide diagnoses or patient-specific treatment or dosing recommendations.",
    "title": "ProtoCite API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/dev-login": {
      "post": {
        "operationId": "dev_login_api_v1_auth_dev_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dev Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "me_api_v1_auth_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refresh_api_v1_auth_refresh_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/feedback": {
      "post": {
        "operationId": "create_feedback_api_v1_feedback_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Feedback",
        "tags": [
          "feedback"
        ]
      }
    },
    "/api/v1/feedback/inbox": {
      "get": {
        "operationId": "inbox_api_v1_feedback_inbox_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_resolved",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Resolved",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedbackOut"
                  },
                  "title": "Response Inbox Api V1 Feedback Inbox Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inbox",
        "tags": [
          "feedback"
        ]
      }
    },
    "/api/v1/feedback/{feedback_id}": {
      "patch": {
        "operationId": "update_feedback_api_v1_feedback__feedback_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "feedback_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Feedback Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Feedback",
        "tags": [
          "feedback"
        ]
      }
    },
...
  }
}